+++ src/api/endpoints/orders/get-orders-bids/v1.ts
@@ -110,7 +110,7 @@ export const getOrdersBidsV1Options: RouteOptions = {
             .items(
               Joi.object({
                 kind: Joi.string(),
-                recipient: Joi.string().lowercase().pattern(regex.address).allow(null),
+                recipient: Joi.string().allow("", null),
                 bps: Joi.number(),
               })
             )