+++ src/migrations/1659596160207_quixotic-order-kind.sql
@@ -1,5 +1,6 @@
 -- Up Migration
 
 ALTER TYPE "order_kind_t" ADD VALUE 'quixotic';
+ALTER TYPE "order_kind_t" ADD VALUE 'nouns';
 
 -- Down Migration